+++ trunk/sim/buffer.cpp
@@ -47,17 +47,17 @@ Tlv::Tlv(unsigned short num, unsigned short size, const char *data)
 {
     m_data.resize(m_nSize + 1);
     memcpy(m_data.data(), data, m_nSize);
-    m_data[m_nSize] = 0;
+    m_data[(int)m_nSize] = 0;
 }
 
-Tlv::operator unsigned short ()
+Tlv::operator uint16_t ()
 {
-    return (unsigned short)((m_nSize >= 2) ? htons(*((unsigned short*)m_data.data())) : 0);
+    return (m_nSize >= 2) ? htons(*((uint16_t*)m_data.data())) : 0;
 }
 
-Tlv::operator unsigned long ()
+Tlv::operator uint32_t ()
 {
-    return (m_nSize >= 4) ? htonl(*((unsigned long*)m_data.data())) : 0;
+    return (m_nSize >= 4) ? htonl(*((uint32_t*)m_data.data())) : 0;
 }
 
 // TlvList
